encryption.py
import base64
import nacl.utils
from nacl.public import PrivateKey, PublicKey, Box
from nacl.encoding import Base64Encoder

def generate_keypair():
    """Generate a new keypair for encryption."""
    private_key = PrivateKey.generate()
    public_key = private_key.public_key
    
    # Encode keys to base64 for storage/transmission
    private_key_b64 = base64.b64encode(private_key.encode()).decode('utf-8')
    public_key_b64 = base64.b64encode(public_key.encode()).decode('utf-8')
    
    return {
        'private_key': private_key_b64,
        'public_key': public_key_b64
    }

def encrypt_message(sender_private_key_b64, recipient_public_key_b64, message):
    """Encrypt a message using the recipient's public key."""
    try:
        # Decode keys from base64
        sender_private_key = PrivateKey(base64.b64decode(sender_private_key_b64))
        recipient_public_key = PublicKey(base64.b64decode(recipient_public_key_b64))
        
        # Create a Box for encryption
        box = Box(sender_private_key, recipient_public_key)
        
        # Encrypt the message
        encrypted = box.encrypt(message.encode('utf-8'), encoder=Base64Encoder)
        
        return base64.b64encode(encrypted).decode('utf-8')
    except Exception as e:
        print(f"Error encrypting message: {e}")
        return None

def decrypt_message(recipient_private_key_b64, sender_public_key_b64, encrypted_message_b64):
    """Decrypt a message using the recipient's private key."""
    try:
        # Decode keys from base64
        recipient_private_key = PrivateKey(base64.b64decode(recipient_private_key_b64))
        sender_public_key = PublicKey(base64.b64decode(sender_public_key_b64))
        
        # Create a Box for decryption
        box = Box(recipient_private_key, sender_public_key)
        
        # Decrypt the message
        encrypted = base64.b64decode(encrypted_message_b64)
        decrypted = box.decrypt(encrypted, encoder=Base64Encoder)
        
        return decrypted.decode('utf-8')
    except Exception as e:
        print(f"Error decrypting message: {e}")
        return None

def verify_encryption_key(public_key_b64):
    """Verify that a public key is valid."""
    try:
        # Try to decode the public key
        public_key = PublicKey(base64.b64decode(public_key_b64))
        return True
    except Exception as e:
        print(f"Error verifying encryption key: {e}")
        
        # In development mode, allow for testing
        import os
        if os.environ.get('FLASK_ENV') == 'development':
            print("Development mode: allowing encryption key verification to pass")
            return True
        return False
