requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
web3==6.12.0
ipfshttpclient==0.7.0
pynacl==1.5.0
psycopg2-binary==2.9.9